Adds compiled TMB code to package
Thank you to @puntae for providing the TMB code and @melissahaltuch-NOAA
for letting me know that it existed. I used pbs-assess/sdmTMB as an
example for how to compile an R package with TMB code. So, thanks to
@ericward-noaa and @seananderson for making that repository easy to
follow. This commit is not complete and will continue to be worked on.